can run loadgens simultaneously

+++ sighandler.c
@@ -1,9 +1,9 @@
 #include "sighandler.h"
 
 struct timespec 	start;
-int 			*pid;
+int 			*pid, *io_pid;
 double 			load, mem_load, io_load, io_max;
-int 			cpu_count;
+int 			cpu_count, do_cpu=0, do_mem=0, do_io=0;
 unsigned long long	mem_size;
 char			*mem_buffer, *io_buffer;
 FILE			*file;
@@ -24,37 +24,35 @@ void sigusr_handler(void){
 
 
 void sigterm_handler(void){
-	int i;
-	for(i=0; i< cpu_count; i++){
-		kill(pid[i], SIGUSR2);
+
+	if(do_cpu){
+		int i;
+		for(i=0; i< cpu_count; i++){
+			kill(pid[i], SIGUSR2);
+		}
+		free(pid);
 	}
-	
-	free(pid);
+
+	if(do_mem){
+		munlock(mem_buffer, mem_size);
+		free(mem_buffer);
+	}
+
+	if(do_io){
+		kill(*io_pid, SIGUSR2);
+		free(io_pid);
+	}
 	exit(0);
 }	
 
-void sigterm_io_handler(void){
-	kill(pid[0], SIGUSR2);
-	
-	free(pid);
-	exit(0);
-}	
-
-void sigterm_mem_handler(void){
-	
-	munlock(mem_buffer, mem_size);
-	free(mem_buffer);
-	exit(0);
-}	
 
 /* exit handler */
 void exit_handler(void){
-	printf("exit\n");
 	exit(0);
 }
 
 void exit_io_handler(void){
-	printf("exit\n");
+	unlink("/home/troy/tmp");
 	fclose(file);
 	free(io_buffer);
 	exit(0);
